+++ bayfox/__init__.py
@@ -1,0 +1,3 @@
+from bayfox.predict import predict_seatemp, predict_d18oc
+from bayfox.plot import predictplot
+from bayfox.utils import get_example_data