+++ bayfox/modelparams/__init__.py
@@ -1,0 +1,1 @@
+from bayfox.modelparams.core import get_draws