+++ deloxfox/__init__.py
@@ -1,3 +1,0 @@
-from deloxfox.predict import predict_seatemp, predict_d18oc
-from deloxfox.plot import predictplot
-from deloxfox.utils import get_example_data