+++ NEWS.md
@@ -1,9 +1,10 @@
-# deloxfox v0.0.1a1
+# bayfox v0.0.1a1
 
+* Rename from `deloxfox` to `bayfox`.
 * Foram 'N_pachyderma_dextral' is now 'N_incompta'.
 * Fixed bug in `predictplot()` caused by bad default argument.
 
 
-# deloxfox v0.0.1a0
+# bayfox v0.0.1a0
 
 * Initial release.